requirements.txt
fastapi==0.92.0
k-means-constrained==0.7.2
openrouteservice==2.3.3
psycopg2-binary==2.9.5
SQLAlchemy==2.0.3
streamlit==1.18.0
streamlit-folium==0.11.1
uvicorn==0.20.0